Refactor JSON reader/writer classes using JSON converter classes
